Extracted RootAttributeDelta to the separate file.
